Split Manual contents into additional pages
New pages: Privacy Policy (with license and credits), Appendix, and

Link the new pages in the Footer; adjusted text size and spacing in the footer.
Updated spacing in the Manual nav panel

Using hidden h1 tags for page titles (verify with @hmarcks on best practice)

Moved the Header component outside of individual pages into <App>.

Replaced Link (react) components in the Header with <a> because Link doesn't work outside the react router.
