+++ src/frontend/src/components/App/index.tsx
@@ -1,11 +1,13 @@
 import React from "react";
 import AppRouter from "../AppRouter";
 import Footer from "../Footer";
+import Header from "../Header";
 
 class App extends React.Component {
   public render() {
     return (
       <>
+        <Header />
         <AppRouter />
         <Footer />
       </>
