+++ src/frontend/src/components/AppRouter/index.tsx
@@ -6,6 +6,9 @@ import RecordSearch from "../RecordSearch";
 import OeciLogin from "../OeciLogin";
 import Landing from "../Landing";
 import Manual from "../Manual";
+import Faq from "../Faq";
+import Appendix from "../Appendix";
+import PrivacyPolicy from "../PrivacyPolicy";
 
 class AppRouter extends React.Component {
   public redirect = () => <Redirect to="/" />;
@@ -18,6 +21,9 @@ class AppRouter extends React.Component {
           <Route component={OeciLogin} path="/oeci" />
           <Route component={RecordSearch} path="/record-search" />
           <Route component={Manual} path="/manual" />
+          <Route component={Faq} path="/faq" />
+          <Route component={Appendix} path="/appendix" />
+          <Route component={PrivacyPolicy} path="/privacy-policy" />
 
           {/* This route catches undefined url entries and redirects back to app
               we could put a 404 page in place at some point */}
