--- src/frontend/src/components/TechnicalDifficulties/index.tsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-class TechnicalDifficulties extends React.Component {
-  render() {
-    return (
-      <p className="bg-white mv4 pa4 br3 fw6 tc shadow br3">
-        Something went wrong, please contact system administrator to report
-        technical difficulties.
-      </p>
-    );
-  }
-}
-
-export default TechnicalDifficulties;
